Conflicts:
	Projects/MacOSFramework/Framework.xcodeproj/project.pbxproj
	Projects/UnitTests/Classes/GameCore.cpp
	Projects/UnitTests/TemplateProjectMacOS.xcodeproj/project.pbxproj
	Projects/UnitTests/TemplateProjectiPhone.xcodeproj/project.pbxproj
	Projects/Win32Framework/Win32Framework.vcxproj
	Projects/iPhoneFramework/Framework.xcodeproj/project.pbxproj
	Sources/Internal/Render/2D/TextBlock.cpp
	Sources/Internal/UI/UIStaticText.cpp
	Tools/ResourceEditor/Classes/Qt/Scene/System/StructureSystem.cpp

+++ Sources/Internal/Render/2D/TextBlock.h
@@ -108,7 +108,7 @@ protected:
 	void Prepare();
 	void PrepareInternal(BaseObject * caller, void * param, void *callerData);
 	
-	void DrawToBuffer(Font *font, int16 *buf);
+	void DrawToBuffer(Font *font, uint8 *buf);
 
 	void ProcessAlign();
